Don't call rx.next after receiving None
In case the channel is dropped, rx.next will return None once.
After that we must not call rx.next or else it will remind us
by panicking. Instead we leave now the restart loop, since
the channel is broken and restarting the actor won't help it.
The supervisor calls a last time .stopped on the actor and
supervising ends.

Unfortunately the doc test does not catch this errornous behaviour
since the panic happens in another thread. The main process still
exits with code 0. We only see the panic through lines on the
terminal.